+++ examples/named_state_with_reference_data.rs
@@ -23,8 +23,8 @@ statemachine! {
 pub struct Context;
 
 impl StatesWithRefDataStateMachineContext for Context {
-    fn action<'a>(&mut self) -> MyStateData<'a> {
-        MyStateData(&42)
+    fn action<'a>(&mut self) -> Result<MyStateData<'a>, ()> {
+        Ok(MyStateData(&42))
     }
 }