+++ CHANGELOG.md
@@ -30,6 +30,7 @@ process.
 - `StateMachine::new` and `StateMachine::new_with_state` are now const functions
 - Fixed clippy warnings
 - [breaking] Changed guard functions return type from Result<(),_> to Result<bool,_>
+- [breaking] Changed action functions return type from () to Result<NextStateData,_>
 
 ## [v0.6.0] - 2022-11-02